+++ R/one_df.R
@@ -165,6 +165,9 @@ drop_k <- function(
   if (inherits(object, "lavaan")) {
     # Ignore sem_out if object is a fit object
     sem_out <- object
+    if (lavaan::lavInspect(sem_out, "fixed.x")) {
+      stop("fixed.x cannot be TRUE for now. Set it to FALSE.")
+    }
     partable <- lavaan::parameterTable(sem_out)
   } else {
     # Assume it is a parameter table
@@ -185,11 +188,14 @@ drop_k <- function(
 
   if (is.null(sem_out)) {
     fixed.x <- partable_fixedx(partable = partable)
+    if (fixed.x) {
+      stop("fixed.x cannot be TRUE for now. Set it to FALSE.")
+    }
     dat <- dummy_data(partable)
     # fit will be used if fit_models is TRUE
     # Need this for lavaan::update()
     fit <- suppressWarnings(
-              do.call(
+              tryCatch(do.call(
                 lavaan::sem,
                 list(
                   model = partable,
@@ -198,7 +204,27 @@ drop_k <- function(
                   fixed.x = fixed.x
                 )
               )
-            )
+            , error = function(e) e))
+    if (inherits(fit, "error")) {
+      if (isTRUE(grepl("not defined in the LISREL representation",
+                      fit$message))) {
+        fit <- suppressWarnings(
+                  tryCatch(do.call(
+                    lavaan::sem,
+                    list(
+                      model = partable,
+                      data = dat,
+                      se = se,
+                      fixed.x = fixed.x,
+                      representation = "RAM"
+                    )
+                  )
+                , error = function(e) e))
+        if (inherits(fit)) {
+          stop(fit)
+        }
+      }
+    }
   } else {
     # Need this for lavaan::update()
     tmp0 <- stats::getCall(sem_out)
@@ -383,6 +409,9 @@ add_k <- function(
   if (inherits(object, "lavaan")) {
     # Ignore sem_out if object is a fit object
     sem_out <- object
+    if (lavaan::lavInspect(sem_out, "fixed.x")) {
+      stop("fixed.x cannot be TRUE for now. Set it to FALSE.")
+    }
     partable <- lavaan::parameterTable(sem_out)
   } else {
     # Assume it is a parameter table
@@ -406,8 +435,11 @@ add_k <- function(
     # fit will be used if fit_models is TRUE
     # Need this for lavaan::update()
     fixed.x <- partable_fixedx(partable)
+    if (fixed.x) {
+      stop("fixed.x cannot be TRUE for now. Set it to FALSE.")
+    }
     fit <- suppressWarnings(
-              do.call(
+              tryCatch(do.call(
                 lavaan::sem,
                 list(
                   model = partable,
@@ -415,8 +447,28 @@ add_k <- function(
                   se = se,
                   fixed.x = fixed.x
                 )
-              )
-            )
+              ),
+              error = function(e) e))
+    if (inherits(fit, "error")) {
+      if (isTRUE(grepl("not defined in the LISREL representation",
+                       fit$message))) {
+        fit <- suppressWarnings(
+                  tryCatch(do.call(
+                    lavaan::sem,
+                    list(
+                      model = partable,
+                      data = dat,
+                      se = se,
+                      fixed.x = fixed.x,
+                      representation = "RAM"
+                    )
+                  ),
+                  error = function(e) e))
+        if (inherits(fit)) {
+          stop(fit)
+        }
+      }
+    }
   } else {
     # Need this for lavaan::update()
     tmp0 <- stats::getCall(sem_out)
@@ -485,12 +537,32 @@ add_k <- function(
   # TODO:
   # - Remove the need to use update
   fit_i <- suppressWarnings(
-            lavaan::update(
+            tryCatch(lavaan::update(
               object = fit,
               model = partable1,
-              warn = FALSE
-            )
+              warn = FALSE,
+              fixed.x = FALSE
+            ),
+            error = function(e) e)
           )
+  if (inherits(fit_i, "error")) {
+    if (isTRUE(grepl("not defined in the LISREL representation",
+                     fit_i$message))) {
+      fit_i <- suppressWarnings(
+                tryCatch(lavaan::update(
+                  object = fit,
+                  model = partable1,
+                  warn = FALSE,
+                  fixed.x = FALSE,
+                  representation = "RAM"
+                ),
+                error = function(e) e)
+              )
+      if (inherits(fit_i, "error")) {
+        stop(fit_i)
+      }
+    }
+  }
 
   # ==== Generate models ====